+++ modules/planning/open_space/distance_approach_problem.cc
@@ -33,7 +33,7 @@ DistanceApproachProblem::DistanceApproachProblem(
 bool DistanceApproachProblem::Solve(
     const Eigen::MatrixXd& x0, const Eigen::MatrixXd& xF,
     const Eigen::MatrixXd& last_time_u, const std::size_t& horizon,
-    const float& ts, const Eigen::MatrixXd& ego, const Eigen::MatrixXd& xWS,
+    const double& ts, const Eigen::MatrixXd& ego, const Eigen::MatrixXd& xWS,
     const Eigen::MatrixXd& uWS, const Eigen::MatrixXd& l_warm_up,
     const Eigen::MatrixXd& n_warm_up, const std::vector<double>& XYbounds,
     const std::size_t& obstacles_num,
